workers/requirements-dynamic.txt
# Optional dynamic-analysis dependencies (install only in isolated lab environments)
# These are intentionally separated from the static baseline requirements.
# Latest verified on PyPI 2026-03-10.
# Note: the PyPI package named `speakeasy` is an unrelated metrics server.
# Use `speakeasy-emulator`, which provides the `speakeasy` malware-emulation module.
# Unicorn still imports `pkg_resources`, so keep setuptools below 81 for now.

speakeasy-emulator==1.5.11
setuptools<81
frida==17.8.0
psutil==7.2.2
